fixed an error in the struct packing for the stram control packet. Now we can direct the client to make any http request, anywhere.
un-broke stream.mp3. It needs to use the new newclient.pm logic. Uses a deviceid of 0 to signifiy an http client.

Now when you play a song from squeezebox, it connects to stream.mp3 and creates a seaprate client, and successfully
plays whatever you tell it to. The next step is to have it request stream.mp3?id=XXXXX, and then HTTP.pm needs to
link it into the existing client instead of creating a separate one.

Also, we have a bug where we try to say hello to previous HTTP clients.... not sure if we did that before too.


git-svn-id: file:///home/awy/mirror/slim/trunk/server@57 62299810-d8cb-41fd-93b7-d32162e5a4a4