Can we agree on the semantics of Client::new with respect to whether the caller does a getClient first or not? Very confusing.
That last change was wrong... was causing us to accumulate squeezebox
clients because every one was hitting client::new in the http handler.
We should never get to Player::HTTP::new for a squeezebox request
(hence the assert I've added).

The actual problem was this typo, which was always returning undef
for getclient because it was looking for the string
"Slim::Player::Client" in the client hash:

<       my $client = Slim::Player::Client->getClient($id);
---
>       my $client = Slim::Player::Client::getClient($id);


git-svn-id: file:///home/awy/mirror/slim/trunk/server@89 62299810-d8cb-41fd-93b7-d32162e5a4a4