Added built-in firmware updater.  Need to bootstrap by manually upgrading to the current version. After this, the updates should be automatically discovered and announced to the user (who will have to press and hold the BRIGHTNESS button for about 5 seconds.
Removed old MP3 CPAN libraries we weren't using.

Merged in a couple of minor changes from the Sourceforge tree.

Renamed the client method "ipaddress" to "ipport" because it returns an ip address
and port in a single string like this:

10.0.1.92:1123

added methods ip and port to just get the part you care about.

Default all transcoding to 320kbps (I really am working on the new transcoding
framework, really...)

Added d_source to see msgs about the audio source processing, indepenedent of the
playlist processing.

Added a whole mess of mac file types that might be found in the wild.

Found a couple of unhandled buttons for use while blocked.


git-svn-id: file:///home/awy/mirror/slim/trunk/server@135 62299810-d8cb-41fd-93b7-d32162e5a4a4