Added patch submitted by Jason Holzapple to allow priority setting from command line.
Tweaked layout of default HTML interface.

Added missing file from handheld skin.

Added contributed Flac library. Need to hook it up still.

Fixed a bug that required rebooting in order to clear the case cache if you
changed the list of non-sorting articles.

Only remember to announce yourself to SLIMP3 players when you start up.
Squeezeboxen should be polling the server.

Fixed a problem where we were getting errors if the player disconnected and
disappeared and we tried to write to closed sockets.


git-svn-id: file:///home/awy/mirror/slim/trunk/server@160 62299810-d8cb-41fd-93b7-d32162e5a4a4