We don't support Mac OS9.  Now, if you ask OSDetect::OS() and it says 'mac' then we're on OSX.
Use this instead of looking for 'darwin' in $^O

Reworked the reworked prefs that KDF made.  Now we try to open the new
preferences path.  If it fails, we try to open one of the old preference
files.  Then we write out the new preferences file every time.  Needs more testing.

Using a suggested tweak to the oggdec parameters to work around a bug in that program.

Added a FAQ about cable length.


git-svn-id: file:///home/awy/mirror/slim/trunk/server@168 62299810-d8cb-41fd-93b7-d32162e5a4a4