Please test this! See if you can stream for a long time. Make sure the interface is always responsive. Test packet loss with large packets, using "ping -s 1400". Thanks!
------

Went through the pcmcia driver from top to bottom and found a couple
bugs and a couple not-really-bugs-but-might-be-a-problem. Changes:

1) Always reset the WE (shared with A0) line before *any* bus cycle - now
that we are talking to the SRAM at interrupt time, it's possible for it
to have changed on us.

2) Set it as early as possible before CE. Previously we would reset it
immediately before asserting CE. Now we do it four cycles earlier
just in case.

Tested with ten different cards, including cards from the 20pc shipment,
both of the new samples we got last week, cards pulled from other devices,
and even a couple prism 2.5 cards I found.

8/10 of the cards worked perfectly.

The other two cards (which had been previously marked as having
problems) had about 10% packet loss. I tested those two cards on the
WET-11 and they exhibited the same problem. I checked signal strength
while streaming and the meter would jump around, seeming to oscillate
between 60% and 90% signal strength. On the working cards it would
always stay at 92% and would only change if I moved the antenna. So
I'm quite sure that those two cards have bad radios, and it is not a
driver or interface problem.


git-svn-id: file:///home/awy/mirror/slim/trunk/server@210 62299810-d8cb-41fd-93b7-d32162e5a4a4