Fixed a couple of warnings messages.
Tweaked the handheld skin a bit.

Added support for raw AIFF and WAV playback.
Works for 44.1k 16bit stereo uncompressed WAV and AIFF files.
By default, transcoding is off for Squeezebox, which is unlikely to work
on a wireless network due to raw data requirements.  Works well on ethernet,
and will work even better when sean pumps up the onboard buffer.

Todo:
        Fix shuffle-on-repeat-playlist, which I broke.
        Automatically turn on transcoding if the player is connected via wireless.
        Add settings to setup area to control the transcoding.
        Fix some cosmetic junk shown on the display between tracks that use different streaming formats.
        Add support (client and server) for mono, 8bit and other sample rates.
        Detect and transcode when we find a compressed WAV file.
        I just heard a click, which implies that we aren't calculating the offset or size correctly somewhere.

Includes a firmware update that is needed for WAV.  Recommended.


git-svn-id: file:///home/awy/mirror/slim/trunk/server@211 62299810-d8cb-41fd-93b7-d32162e5a4a4