Added a 4th parameter to convert.conf that is the individual player ID.  This change is NOT backwards compatible with the previous version.  If you wish to use your old convert.conf, you need to add the 4th parameter as '*'.
Currently the ID is based on the MAC address of the player if it's a
slim device, and the IP if it's an HTTP connection.  This could be a
problem because:

    * Your computer may be assigned a dynamic IP address.
    * Your computer may be behind a NAT firewall, in which case all
    IPs from withing the firewall will be the external address of the
    firewall.

Perhaps the ID for the purpose of this parameter should simply be the
MAC address.  This would be globally unique, right?  However, IP based
would allow you to say, "all users from outside my firewall get the
bitrate-reducing filter."  Perhaps a 5 parameter setup is needed, with
IP and MAC with full support for widlcards.


git-svn-id: file:///home/awy/mirror/slim/trunk/server@225 62299810-d8cb-41fd-93b7-d32162e5a4a4