The Getting Started document will have the changelog at the end.
We are "Updating" the software, not "Upgrading"

Updated the firmware image to handle the IP address of the server changing on the same subnet.
Makes things more reliable on ad hoc networks.

I got a crash that I couldn't reproduce.  Putting in a sanity check that might prevent this.

Put some additional fields in the XML status for extraction by another program.


git-svn-id: file:///home/awy/mirror/slim/trunk/server@274 62299810-d8cb-41fd-93b7-d32162e5a4a4