Rework of select framework.  Not a substantial win in terms of performance, but we do call select() a hell of a lot less often.
Use Slim::Networking::Select's functions to register a socket that you'd like
a callback on when it's ready to read or write.

Updated some documentation.

Fixed a bug in the firmware that could case a (eek) packet storm.

Fixed the xml skin so that it works in IE.

Fixed a problem where we were sometimes not skipping onto the next song.


git-svn-id: file:///home/awy/mirror/slim/trunk/server@275 62299810-d8cb-41fd-93b7-d32162e5a4a4