Add wildcard handling for button pressing styles. If a button is followed by .* in the mapping file, the .* will be expanded into all of the pressing styles (enumerated in @buttonPressStyles).
If a particular style follows the wildcard in the mapping file, that mapping
will override the one assigned by the wildcard.  If the wildcard comes last,
it will override the previously assigned mapping.

This is primarily useful when killing off mappings (assigning the button to
the dead function) to make sure that future changes won't result in unexpected
behavior.


git-svn-id: file:///home/awy/mirror/slim/trunk/server@278 62299810-d8cb-41fd-93b7-d32162e5a4a4