Finally got around to fixing up the Windows installer.
Now it asks for your user password on NT and above so that it can run as you.  This makes it possible to access iTunes libraries.

Droo should be happier now that the uninstaller cleans up better after itself.

Had to tweak the service code in the server to be able to set itself up under a given user.

Had to tweak the win app code because apparently if you start a windows service as a particular user, you can't get information about that process
even if you are that user.  And they say Palladium will make our lives easier...


git-svn-id: file:///home/awy/mirror/slim/trunk/server@315 62299810-d8cb-41fd-93b7-d32162e5a4a4