implemented numberScroll for INPUT.List Added isSorted parameter, which is used by the numberScroll function to set up the call to Slim::Buttons::Common::numberScroll. If isSorted is false, then the listRef parameter will be sent to numberScroll and the sorted parameter of that function will be set to 0.
If isSorted is 'e' or 'E', the externRef parameter will be sent to numberScroll
and the sorted parameter of that function will be set to 1.

If isSorted is 'i' or 'I', the listRef parameter will be sent to numberScroll
and the sorted parameter or that function will be set to 1.


Added default actions for exit_left and exit_right of popModeRight and
bumpRight respectively.  This allows the standard behavior to cover simple
list display without needing a callback function.


git-svn-id: file:///home/awy/mirror/slim/trunk/server@345 62299810-d8cb-41fd-93b7-d32162e5a4a4