Make it so we handle odd-size segments correctly.
No memory realignment is done. Instead, we save each segment into the buffer with a
size header. Then at interrupt time, when reading out of the buffer, we have a
special case at the end of an odd-sized segment - we just send one byte to the
serdes.

This does have the effect of making the fullness calculation off by approximately
0.1% due to the size header, but I don't see how that could be a problem.

Internet radio plays okay now. Also tested with files at various bit rates. This
still needs to be tested on a stable server for a day or so of continous streaming.

Firmware updater bug has not recurred. Tested going forward and back to/from a
two-week old version of the firmware, and it's okay. Pretty sure it was a
compilation bug on my system at home.

This is ready for testing. Please install!


git-svn-id: file:///home/awy/mirror/slim/trunk/server@358 62299810-d8cb-41fd-93b7-d32162e5a4a4