Bumping pre-release version number to 5.0.3
Caleb Epstein fixed a problem with cue parsing.

Fixed a problem where we'd often say Disk 1 of 1.  That's unnessary.

Distinguish in the debugging output between when a squeezebox disconnects and
when an HTTP client disconnects.

Make sure that the currentsongindex is reset to zero appropriately to fix a warning.

After a forced upgrade in squeezebox, unblock the client to avoid a glitch.

Added a patch that fixes a problem with the Screensaver.

Victor submitted a fix for the RPM to make it complain less during installs.  Fixes an old bug.


git-svn-id: file:///home/awy/mirror/slim/trunk/server@389 62299810-d8cb-41fd-93b7-d32162e5a4a4