adding supprt for lyric display, start animation with custom pause via a client pref.  this will have to be set back to default after a lyric call that alters it.
web settings player pref to aler this pause directly, so it can be
tested

this hasn't hurt my system, but I know animation might be tricky with
some.  Feel free to send emails my way if this causes problems.  I do
plan to rework this on the whole eventually, but the first plan is to
support lyrics.


git-svn-id: file:///home/awy/mirror/slim/trunk/server@425 62299810-d8cb-41fd-93b7-d32162e5a4a4