Tim Long pointed out some inconsistencies in the CLI documentation and API. This was due to our slow migration away from IP addresses as player identifiers toward unique MAC identifiers. The code and documentation has been updated to this: Use this command:
player id <n>

to obtain the unique identifier for a player, which may be in the form of a MAC address or IP address or something else in the future.

then use the returned <id> to control the player:

<id> pause

you can still get the IP address for a given player:

player ip <n>

And the name, of course:

player name <n>

And these should also work as identifiers for control.  Note that IP addresses and port numbers may change over time, so they aren't recommended for use as identifiers.


git-svn-id: file:///home/awy/mirror/slim/trunk/server@461 62299810-d8cb-41fd-93b7-d32162e5a4a4