fix bug 84 and add a player web setting for the built in server alarm clock
move towards a fix for bug 35, but adding the option to use variables
for the cover art.  prefix the line with % and you can use a title
Format style line instead of a filename.
eg %ARTIST - TITLE, would give U2 - Joshua Tree.jpg when the ID3 tags
were Artist = U2 and Album = Joshua Tree.  The strings might need to
be improved to explain this more clearly.

I'm looking for input on this.  This is currently the simplest way to
add variables and still make use of existing code.

Full fix of bug 35 would need to add the ability to choose a cover art
folder.


git-svn-id: file:///home/awy/mirror/slim/trunk/server@467 62299810-d8cb-41fd-93b7-d32162e5a4a4