v.6 firmware with new flow control mechanism to fix mp2 streams
tested extensively going back and forth to the previous CVS version and
the v5.0.1 release (which both had a v.5 version #). No problem with the
update process.

However, in the course of testing different firmware revs I did find that
a new bug may have crept in since 5.0.1, with repect to PCM/MP3
switch-over. Sometimes (about 1/20) when going from mp3 to PCM, the new
stream does not start. Looking for a race condition during decoder
initialization. I am sure however that this bug is not related to the new
flow control because the behavior is exhibited also in the previous rev.
Reviewing all the changes since 5.0.1.


git-svn-id: file:///home/awy/mirror/slim/trunk/server@500 62299810-d8cb-41fd-93b7-d32162e5a4a4