Fix a warning when sorting tracks that don't have track numbers.
Fix a problem in the previous fix for 404 not found errors.  If, in generateResponse, the
reference to the body is undefined, then we've started working on the page in the background.
If the reference points to an undefined value, then we were unable to find the data, and thus, 404 Not Found.


git-svn-id: file:///home/awy/mirror/slim/trunk/server@531 62299810-d8cb-41fd-93b7-d32162e5a4a4