Upcoming firmware revisions will use a new update mechanism. We're going to jump to from rev 8 to 20 for the new SDK5-based firmware and that'll be how we tell the difference for the purpose of loading a firmware image.
Instead of the server connecting to the squeezebox and uploading the
image, the new mechanism will use slimproto. This means, among other
things, that the control channel stays connected, and we can display a
nice status screen during the update. Some changes to slimproto:

Deprecated client->server opcode: BYE!. We might use this later for
something else. In the mean time, it is still of course supported for
older firmware revs.


New client->server opcodes:

	UPRQ	update requested

New server->client opcodes:
	upda	"update data" chunk
	updn	"update done" - signals end of update image


git-svn-id: file:///home/awy/mirror/slim/trunk/server@697 62299810-d8cb-41fd-93b7-d32162e5a4a4