d_protocol_verbose only used in two places. Squeezebox.pm can use d_slimproto_v and the other can be removed since slimserver.pl doesn't support the d_protoco_verbose flag.
fix alignment of usage message in slimserver.pl

When we are skipping through songs because they are not found, lets
put up a message on the player display and wait 1/4s before trying the
next song.  This gives the user an idea of what is happening instead
of looking like its stuck on stop.  1/4s shoudl be short enough so
that you don't have to wait too long before you get to a good song.


git-svn-id: file:///home/awy/mirror/slim/trunk/server@705 62299810-d8cb-41fd-93b7-d32162e5a4a4