Added settings page for enabling/disabling convert.conf entries.  Useful if you want to force LAME transcoding if you have a slower network.
Also, we now check to see if the appropriate binaries are available and skip
any configurations that require specific binaries.  The syntax of the convert.conf file has
been changed slightly, now binaries are expressed as [binaryname] instead of $binaryname$

Older config files that use the $binaryname$ format should still work, but won't get run-time checks for the presence of the binaries.

Squeezebox now also "prefers" MP3 conversion when possible.

All of this is to make life easier for new customers who are trying to get AAC and other non-MP3 formats working with the least amount of messing.
Out of the box, the behavior is similar as it was, but if you install LAME, it will try to use it unless you disable that line in the configuration.


git-svn-id: file:///home/awy/mirror/slim/trunk/server@720 62299810-d8cb-41fd-93b7-d32162e5a4a4