Added a commented out line in HTTP that may get re-enabled if we find that some platforms allow the setsockopt for the send buffer to be larger than the receivers TCP window size.
Accepted Peter Speck's patch for fixing the bug in the regexp in transcoding that failed to detect missing binaries used in transcoding.
Also, Peter patched up some problems in the AIFF parser:

The sample rate is an 'extended', which is 10 bytes. Only first 6 bytes was read, so $encoding was always garbage.

Chunks are always an even number of bytes, even if the chunkSize entry is odd (it is null-byte padded).  All my AIFF files have an even chunk size for all chunks except the last chunk, which is an ID3 entry lacking the null-padding. So it have not been a problem for me, just noticed it...

Removed the redundant tell($f) statements and made the chunk-seeking logic a simple while() loop, which makes the diff big.
I have renamed chunkoffset to dataoffset, as it is the offset of the sound data, and not the offset of the chunk.

Updated the convert.conf file so that when LAME is invoked to encode into MP3, we force to resample at a 44.1k sample rate.  Some non-standard sample rates were causing LAME to fail.

Kevin Pearsall found that convert.conf file had some extraneous "1" options in the oggdec entries, which breaks under the latest oggdec.


git-svn-id: file:///home/awy/mirror/slim/trunk/server@729 62299810-d8cb-41fd-93b7-d32162e5a4a4