Changes to the upgrade mechanism in order to support upgrading to the new SDK5-based firmware:
All versions < 10 are forced to 10

New version 10 (binary checked in) can be upgraded to 20. This is the ONLY
change from v8 to v10.

squeezebox.version now specifies the upgrade path for every firmware
revision.

Update using firmware.html always upgrades to version 10 using the old
mechanism.

If an upgrade image is specified in squeezebox.version but the upgrade
file does not exist, the server prints a friendly notice saying so, but
does not force an upgrade.

This has been tested successfully for upgrading from v8 to v10 to v20. v20
is not included here yet, but will be ready for beta testing in a few
days.

Dean: this is exactly what we discussed except that we don't list the
explicit filename in squeezebox.version. The reason for that is because
the client doesn't specify what filename he's running - he says what
version number. So if you press & hold brightness when no update is
forced, the server doesn't know what filename to give him corresponding to
the version he's already running. I decided it was easier to just stick to
one simple naming convention that the server knows, instead of trying to
be excessively clever about it in squeezebox.version.


git-svn-id: file:///home/awy/mirror/slim/trunk/server@737 62299810-d8cb-41fd-93b7-d32162e5a4a4