Phil Barrett suggested a patch for bug 219, to make the # sign more attractive when in double size.
New firmware will report buffer size and fullness rather than read and write pointer.  Account for this in the server.

Working on fixing some problems with FFWD and REW trick modes.  i

Seems better about jumping to random locations, but occasionally still won't go back to 1x when we're towards the end of a song.
Working on this too.


git-svn-id: file:///home/awy/mirror/slim/trunk/server@753 62299810-d8cb-41fd-93b7-d32162e5a4a4