Added a server setting to allow folks to set the minimum wait time between iTunes library importing.  The default is (and was) 60.
Peter Timofejew submitted a patch to not import songs marked as disabled in iTunes.

Peter Timofejew submitted a patch to fix a problem with automount paths in iTunes XML files.

Pat Farrell updated the unix Installation.txt file.

Rob Funk submitted a patch that ignores punctuation for the purposes of sorting.


git-svn-id: file:///home/awy/mirror/slim/trunk/server@771 62299810-d8cb-41fd-93b7-d32162e5a4a4