There was a conflict between the parameter used to declare a page title and the parameter used to declare a song title. Changed the page title to "pagetitle" from just "title"
Added a prototype pitch control.  It only works on Squeezebox now and only with MP3 files and only after a song is restarted.  But it's a start.

Updated patch for AIFF files from Caleb Epstein

Accepted patch that reduce custom character flicker from Phil Barrett

Another tweak to the bitrate rounding.

Trying to reduce latency in the HTTP streaming.  Clearing the buffer if things change.


git-svn-id: file:///home/awy/mirror/slim/trunk/server@794 62299810-d8cb-41fd-93b7-d32162e5a4a4