reverting EN/status_list.html since it wasn't a real fix.
in Pages.pm, building status.html will pass artist and album in the
params hash to playlist.  This will cause the playlist to think there
is artist info when there shouldn't be.  In the case where artist and
album are included in titleformat, it should be undef, so to be safe
lets make sure its undef for the playlist in this case.


git-svn-id: file:///home/awy/mirror/slim/trunk/server@809 62299810-d8cb-41fd-93b7-d32162e5a4a4