KDF fixed some problems with Info.pm
On startup, if using the tag database only, the artwork is missing.  I've added
the artworkCache to the stored db to fix this, and uppped the DBVERSION.

using iTunes, the song count is 0 and this is because the fileURL doens't pass
the -e test.  I've added a case to allow it to pass when using iTunes, and to
count new songs.


git-svn-id: file:///home/awy/mirror/slim/trunk/server@828 62299810-d8cb-41fd-93b7-d32162e5a4a4