Update from Peter for SHOUTcast browser.  Provides a nice way to customize the Genres you see.
Patch from Richard Titmuss to help us get SoftSqueeze served as a Java web launch application.

Added "unuri" filter to remove URI escaping for human readability. Use this when displaying the path to a song in songinfo

A few itunes lines were missing from a patch that Richard Purdie sent.  Updated.

When we parse the strings file, uppercase the language codes so we don't get errors on trivial typos.


git-svn-id: file:///home/awy/mirror/slim/trunk/server@861 62299810-d8cb-41fd-93b7-d32162e5a4a4