Reworked the rework of fileURLFromPath to let URI::file fill out the url scheme.
Removed the reference loop in the HTTP params.  Fix a couple of templates to handle this.

Add a warning if we try to get a path from a file URL if it's not a file URL.

Fixed some miswords in the warnings when we pass a path instead of an expected URL.


git-svn-id: file:///home/awy/mirror/slim/trunk/server@878 62299810-d8cb-41fd-93b7-d32162e5a4a4