Add $ID tag to Client.pm
Letting users decide the output format to their clients is a good
idea.  This can be done using max bitrate, henceforth known as bitrate
limiting.

This player setting can be used to control the bitrate of mp3 sent to
http/slimp3 clients.  MP3 files already below the bitrate will be sent
unaffected.  For clients that can handle PCM, there is an option for
"no limit", regardless of wired/wireless connection, the 'no limit'
will convert all non-mp3 formats to PCM.  MP3 formats are left
unchanged.  setting a bitrate for these clients will have the same
effect as for http/slimp3.

Default is 320, but it would be nice to have a clean way of defaulting
wired squeezeboxes to 'no limit'.  I'm thinking checkClientPrefs, but
would love to know if there are nicer ideas.

A request is filed for better feedback, so have tried to do a few
things.  Try this stuff out, and off up any ideas for better feedback
if you think it is needed.


git-svn-id: file:///home/awy/mirror/slim/trunk/server@922 62299810-d8cb-41fd-93b7-d32162e5a4a4