Large patch from KDF & RP:
* Default wired squeezebox to no bandwidth limit.

* Slimserver will no longer issue a message if audiodir is blank, so long as either moodlogic or itunes import are in use.

**A patch from Richard does several things:

* Adds a TTL (Time to Live) field. This is used to see when the file was
last checked for changes. Every 5 minutes, it will compare filesize and
timestamp to see if it's changed. We don't do this on every access to the
Cache as perl IO routines to do this are slow on windows. This means we
should finally notice files that change and files that are deleted (within 5
mins of a change).
* Make all cache data accessible whilst rescans take place. Tricky but it
does this now. (The 'VALID' field is used to check whether data has been
touched as part of the rescan. If the data hasn't been marked as valid, it
gets wiped. This only happens upon rescan though).
* Allows all the data sources to be used at the same time (Moodlogic,
iTunes, MusicFolder).  The rescan structure has been adjusted to mean all
enabled data sources get rescanned upon a rescan. To do this I've added an
extra file Slim::Music::Scan which controls MoodLogic.pm, iTunes.pm and
MusicFolderScan. Any call to rescan should now go to
Slim::Music::Scan::startScan which should do all the work
(Music::MusicFolderScan and Utils::Scan could probably be merged now).
* Fix iTunes code which was importing file://localhost/ instead of file:///
leading to duplicate entries in the cache
* Fix several song count errors. All these changes should leave the song
counters much more robust
* Fix some artwork code that was triggering a backtrace (fileURL conversion
leftover)
* Get rid of some unneccessary calls to CacheEntry before calling
updateCacheEntry (it will happily merge data itself)

* Settings and Strings altered as required for this patch.

* fixed a bad debug tag in itunes.pm

* Alter page generation so that "music folder" link does not appear if audiodir is not set.
* explicitly create the music folder link in home.html using addLinks, if audiodir is defined.
	- should probably move towards generating all the home links this way.

* add 'item' paramref for use with browseid3_artwork.html, and future ability to have artork in other browse modes.


git-svn-id: file:///home/awy/mirror/slim/trunk/server@942 62299810-d8cb-41fd-93b7-d32162e5a4a4