Fix some problems with the home.html pages on various skins now that we insert many of the links dynamically.
Some of the home.html pages were indistinguishable from the default EN skin. Removed them.
(Olson skin was the exception, but that's worked around by redirecting browseid3.html to olsonmain.html.)

Kevin Pearsall noticed that web browsers were caching pages too agressively even after we change skins.  Now when we first load an item into the html
cache, we return a modification time of time() to force expiration of client caches after reload.

Decide on whether or not to show the Music Folder on the Squeezebox Home screen based on whether we have one defined.
Show Saved Playlists directory if the user has a playlist directory defined or they are using itunes or moodlogic.

Did some work on fixing the problem of some receivers not syncing up properly at the beginning of tracks.
 The silence mp3 data that I had created had the wrong sample rate.
 Now when we scan through a file or try to jump to the middle, try to line up on MP3 frame boundries.
 This should keep the S/PDIF receivers from dropping out.


git-svn-id: file:///home/awy/mirror/slim/trunk/server@957 62299810-d8cb-41fd-93b7-d32162e5a4a4