Bug291: Show PCM/WAV info in UI.  Output bitrate is now shown in Web Status pane. Not yet added to player UI, but could be in trackinfo.
Bug313 - Wireless signal strength is now shown at the bottom of the
player settings page along with MAC, IP, firmware.

Some fixup of the transcoding conversion profile.  We hide it from
users, so it shouldn't ever be disabled.  Provide feedback in player
settings for Bitrate Limiting to show whether or not we've found the
lame executable. This should be a live thing, so if lame isn't found,
but then later installed, we'll say so and work with, without
requiring a restart.


git-svn-id: file:///home/awy/mirror/slim/trunk/server@965 62299810-d8cb-41fd-93b7-d32162e5a4a4