Fixed a problem where the spinning cursor animation while blocked on long operations (like entering a LARGE directory) was broken.
Firmware 40:
  - Fix a problem with 32 character long SSIDs.
  - When reconnecting, the player tells the server if it rebooted.  The server will now only restart streams if the player has rebooted.

Added postulated fix for WMA files that have broken headers with block lengths that are shorter than possible.

Contributed patch to workaround bogus modification dates on folders that are the root of windows volumes.

Removed scantime parameter that wasn't calculated properly when starting scans.  Use the log timestamps if you want to know how long.

Patch from KDF to fix problem with double-animations when moving in to browse music or playlists.


git-svn-id: file:///home/awy/mirror/slim/trunk/server@1392 62299810-d8cb-41fd-93b7-d32162e5a4a4