Fix bug652: when browsing a playlist via the player UI, we now move to the next song when a song is deleted.
Fix Bug663: add a setting in player settings to choose the LAME
quality level per player for mp3 conversion.  9 is fastest with ok quality and is the
default.  0 is best quality but uses a lot of cpu power.

Fix bug649: Allow assignment of screensaver in off mode.  We now have
settings for screensavers while playing, while stopped, and when off.
all tests for power on/off state must now use $client->power().
$client->mode eq 'off' is no longer userful for this.


git-svn-id: file:///home/awy/mirror/slim/trunk/server@1485 62299810-d8cb-41fd-93b7-d32162e5a4a4