fix Bug 783: wireless signal strength doesn't auto-update.
This would update when playing music, but not when idle.  This is
because we do not get STAT updates from a SqueezeboxG when idle.  The
simple fix is to insert a blank i2cc frame when we grab signal
strength from the Information module.  The signal strength will now
update every second.



git-svn-id: file:///home/awy/mirror/slim/trunk/server@1763 62299810-d8cb-41fd-93b7-d32162e5a4a4