Add Plugin modes during the init of Slim::Buttons::Plugins rather than during the init of Slim::Buttons::Common.
This makes the adding of web pages happen after Slim::Web::HTTP init, so the hack can be removed.

Make sure that the order of directories for web pages and plugins is ~/Library/Slimdevices/ first, then /Library/Slimdevices, and finally $Bin.

The use lib in Slim::Buttons::Plugins adds the appropriate directories to @INC, so no need to unshift them there as well.


git-svn-id: file:///home/awy/mirror/slim/trunk/server@1944 62299810-d8cb-41fd-93b7-d32162e5a4a4