Merging import_plugin_branch rev 1861:2029
This moves the functions of itunes, MoodLogic and Musicmagic to
plugins so that future importers of thie type will be easier to add.

MusicMagic and Moodlogic provide an optional mixer function.  These
links are generated dynamicly for the browse modes where skins support
this feature.  So far, support is included for EN, Default, Fishbone.  

Server settings still present Use and Dont Use options, which are also
presented in server settings->plugins.  Plugins can also be disabled
which should remove all the importer-related functions for the
disabledd plugins (ie no more MOODLOGIC_MIXABLE requests when the user
doesn't care about moodlogic)

Some hooks are still locked into the coer server code, that should
hopefully migrate to a pluggable API, such as the table fields related
to moodlogic, etc, also the URI constructors and the isMixable
methods.  

So who is going to do Rhythmbox?? ;)



git-svn-id: file:///home/awy/mirror/slim/trunk/server@2030 62299810-d8cb-41fd-93b7-d32162e5a4a4