Schema update to include disc as part of the tracks table.
This will trigger an automatic wipe & rebuild of the database.

Allows a fix to go in for bug #876 - making groupdiscs work again.

Each track should know what disc it belongs to, in the case of a single album object.

Dynamically create the title for multi-disc albums, instead of storing a fixed title in the db.

===

Add some methods to clear the ContributorTrack & GenreTrack caches. Having
stale objects around led to wacky things like 0 Genres and 0 Contributors
showing up after a wipecache.

===

return early from _postCheckAttributes if the "track" is really a directory or
windows shortcut. This gets rid of "No Album", etc if there aren't any.

May need to be adjusted to add playlists as well.



git-svn-id: file:///home/awy/mirror/slim/trunk/server@2253 62299810-d8cb-41fd-93b7-d32162e5a4a4