Fix bug 907.  We weren't staying in the off mode when off, so mappings there weren't being used.
Now screensavers which register themselves with a name starting with "SCREENSAVER." will add a second mode which starts with "OFF."  This results in the mapping for the off mode being used rather than that of the screensaver mode.

Plugins which want to explicitly set the mapping for themselves when used can supply a mapping for this second mode.

Also removed some unnecessary wildcard mappings and made play.single in off mode explicitly call the play function in the off mode, since screensavers will not be likely to provide their own play function which does the right thing.



git-svn-id: file:///home/awy/mirror/slim/trunk/server@2374 62299810-d8cb-41fd-93b7-d32162e5a4a4