Fix CPU spiking - from Adrian:
I think this only catches the case where a timer is set in the future and
HTTP->content() is called from within a timer.  I was calling idleStreams()
with a timeout of 0.1 in this case, but it would be reset to 0 by the next
timer test and then the timer is blocked from executing.



git-svn-id: file:///home/awy/mirror/slim/trunk/server@2467 62299810-d8cb-41fd-93b7-d32162e5a4a4