more fixes for Bug992: wired and wireless SB2 defaults to no limit for bitrate.
MusicMagic Plugin: remote access settings are confusing and only work
for very specific cases, so drop it until a proper solution can be
found.  Leaving it in commented, so that hackers have the option to
still use it.

Fix Bug 1088: don't allow plugins without a mode or functions to be
listed in the Home Menu options as these are NOT player UI plugins
  Users who have already attempted to add these to the home menu, can
  simply remove them to stop the potential crash.  The plugin will
  then drop off the list, into the safe state.



git-svn-id: file:///home/awy/mirror/slim/trunk/server@2521 62299810-d8cb-41fd-93b7-d32162e5a4a4