resolution to bug 1069
fixed fade_volume, which was broken when fading up (and possibly down as well).  Client now stores temporary volume settings, and players take care not to overwrite the temporary setting.  This allows fade_volume to set the volume with the temp flag, and have that setting actually be used.

I've also added a client pref, alarmfadeseconds, which is the amount of time over which the alarm should fade in.  Default is 0 seconds, or no fade - same as current bahavior.  I have not added a convenient way to adjust this setting, a user would have to edit the prefs file.  So most users will not notice any change with this checkin.




 


git-svn-id: file:///home/awy/mirror/slim/trunk/server@2534 62299810-d8cb-41fd-93b7-d32162e5a4a4