fix Bug 1118: importers are not read before Browse::init() is called, so if we have no playlistdir setting, then the playerUI does not allow BrowsePlaylists, even if we do have imported playlists.
Move browse menu to its own function that we can call from
Home::updateMenu so that this can be changed as needed when settings
are changed.  This also revealed a problem where having an audiodir
and not playlistdir would still list Browse Playlist in the web UI.
This is becuase FOLDER is now an import scanner.  Since we use the
countImports() function to tell if we're useing non-folder imports,
not counting the FOLDER scan is the more correct thing to do.
Simplifies the test in slimserver.pl as well for valid data sources.




git-svn-id: file:///home/awy/mirror/slim/trunk/server@2553 62299810-d8cb-41fd-93b7-d32162e5a4a4