use level 1 links for lists.
it doesn't look at first like this does much, but it does.

an example:  a manual search lists songs with album and artist links,
click on an album link and you get a list of browse songs, matching
that album.  with this change, you get the browse album page for that
album (most noticeable difference, the latter includes artwork)

done for fishbone, EN and Default skins



git-svn-id: file:///home/awy/mirror/slim/trunk/server@2562 62299810-d8cb-41fd-93b7-d32162e5a4a4