Fix a memory leak in non-streaming HTTP response.
$response->content() was too messy and obtuse.

Pass $body along as a ref, and add two items onto the httpClient output buffer
- the first for headers, the second for data. This prevents the need to do an
extra in memory copy of the body.

If Scalar::Utils is available, weaken the reference as well to make sure it's cleaned up.

For other places where the content is small (error responses), use the
$response->content_ref() method, instead of dereferencing the return value
from fillintemplate().



git-svn-id: file:///home/awy/mirror/slim/trunk/server@2587 62299810-d8cb-41fd-93b7-d32162e5a4a4