Fix for the "Softsqueeze not working" issue - which really is:
We were resetting the $body reference to "" when a HEAD request was made, as
is done by the JWS application. This was also the cause of the templatefiles
cache returning 0 for static content.

So - Don't reset $body to "" when it's a HEAD (or Not Modified) request.

Also - don't cache the SoftSqueeze files in the first place - 4.7M on disk -
takes up a lot in the cache, when JWS does it's own HEAD checking.



git-svn-id: file:///home/awy/mirror/slim/trunk/server@2602 62299810-d8cb-41fd-93b7-d32162e5a4a4